fixed for compile error

+++ nor_isp.sh
@@ -10,14 +10,20 @@ F=fip.img
 cp $X xboot
 cp $U uboot
 cp $S spi_all
-cp $F fip
+
 # There are 3 partitions in ISPBOOOT.BIN
 # xboot (64k)
 # uboot (64k)
 # spi_all (size of SPI-NOR)
-isp pack_image4nor_isp ISPBOOOT.BIN xboot uboot spi_all 0x1000000 fip 0x100000
+
+if [ "$1" == "Q645" -o "$1" = "SP7350" ]; then
+	cp $F fip
+	isp pack_image4nor_isp ISPBOOOT.BIN xboot uboot spi_all 0x1000000 fip 0x100000
+	rm -f fip
+elif [ "$1" == "Q628" ]; then
+	isp pack_image4nor_isp ISPBOOOT.BIN xboot uboot spi_all 0x1000000
+fi
 
 rm -f xboot
 rm -f uboot
 rm -f spi_all
-rm -f fip
